Modularization and cleanup for PyPi
