+++ onto_tool/__main__.py
@@ -1,0 +1,5 @@
+"""Module entry point."""
+
+if __name__ == '__main__':
+    from .onto_tool import run_tool
+    run_tool()
